move obj loading to own class
